[Core] Moving flag from Condition/Element to GeometricalObject
Basically that. The flag is common between the geometries and conditions, but currently both derive from Flag instaed of bein included on the GeometricalObject
