+++ kratos/includes/element.h
@@ -4,8 +4,8 @@
 //   _|\_\_|  \__,_|\__|\___/ ____/
 //                   Multi-Physics
 //
-//  License:		 BSD License
-//					 Kratos default license: kratos/license.txt
+//  License:         BSD License
+//                     Kratos default license: kratos/license.txt
 //
 //  Main authors:    Pooyan Dadvand
 //
@@ -21,7 +21,6 @@
 #include "includes/properties.h"
 #include "includes/process_info.h"
 #include "includes/geometrical_object.h"
-#include "containers/flags.h"
 #include "includes/constitutive_law.h"
 #include "containers/weak_pointer_vector.h"
 
@@ -57,7 +56,7 @@ namespace Kratos
  * not all of them have to be implemented if they are not needed for
  * the actual problem
  */
-class Element : public GeometricalObject, public Flags
+class Element : public GeometricalObject
 {
 public:
     ///@name Type Definitions
@@ -119,7 +118,6 @@ public:
      */
     explicit Element(IndexType NewId = 0)
         : BaseType(NewId)
-        , Flags()
         , mpProperties(nullptr)
     {
     }
@@ -129,7 +127,6 @@ public:
      */
     Element(IndexType NewId, const NodesArrayType& ThisNodes)
         : BaseType(NewId,GeometryType::Pointer(new GeometryType(ThisNodes)))
-        , Flags()
         , mpProperties(nullptr)
     {
     }
@@ -139,7 +136,6 @@ public:
      */
     Element(IndexType NewId, GeometryType::Pointer pGeometry)
         : BaseType(NewId,pGeometry)
-        , Flags()
         , mpProperties(nullptr)
     {
     }
@@ -149,7 +145,6 @@ public:
      */
     Element(IndexType NewId, GeometryType::Pointer pGeometry, PropertiesType::Pointer pProperties)
         : BaseType(NewId,pGeometry)
-        , Flags()
         , mpProperties(pProperties)
     {
     }
@@ -158,7 +153,6 @@ public:
 
     Element(Element const& rOther)
         : BaseType(rOther)
-        , Flags(rOther)
         , mData(rOther.mData)
         , mpProperties(rOther.mpProperties)
     {
@@ -184,7 +178,6 @@ public:
     Element & operator=(Element const& rOther)
     {
         BaseType::operator=(rOther);
-        Flags::operator =(rOther);
         mpProperties = rOther.mpProperties;
         return *this;
     }
@@ -194,7 +187,7 @@ public:
     ///@{
 
     /** Dimensional space of the element geometry
-	@return SizeType, working space dimension of this geometry.
+    @return SizeType, working space dimension of this geometry.
     */
 
     KRATOS_DEPRECATED_MESSAGE("This is legacy version, please ask directly the Geometry") SizeType WorkingSpaceDimension() const
@@ -427,9 +420,9 @@ public:
                                       ProcessInfo& rCurrentProcessInfo)
     {
        if (rLeftHandSideMatrix.size1() != 0)
-	  rLeftHandSideMatrix.resize(0, 0, false);
+      rLeftHandSideMatrix.resize(0, 0, false);
         if (rRightHandSideVector.size() != 0)
-	  rRightHandSideVector.resize(0, false);
+      rRightHandSideVector.resize(0, false);
     }
 
     /**
@@ -459,7 +452,7 @@ public:
                                        ProcessInfo& rCurrentProcessInfo)
     {
         if (rLeftHandSideMatrix.size1() != 0)
-	  rLeftHandSideMatrix.resize(0, 0, false);
+      rLeftHandSideMatrix.resize(0, 0, false);
     }
 
     /**
@@ -470,8 +463,8 @@ public:
      * @param rLHSVariables parameter describing the expected LHSs
      */
     virtual void CalculateLeftHandSide(std::vector< MatrixType >& rLeftHandSideMatrices,
-					const std::vector< Variable< MatrixType > >& rLHSVariables,
-					ProcessInfo& rCurrentProcessInfo)
+                    const std::vector< Variable< MatrixType > >& rLHSVariables,
+                    ProcessInfo& rCurrentProcessInfo)
     {
     }
 
@@ -485,7 +478,7 @@ public:
                                         ProcessInfo& rCurrentProcessInfo)
     {
         if (rRightHandSideVector.size() != 0)
-	  rRightHandSideVector.resize(0, false);
+      rRightHandSideVector.resize(0, false);
     }
 
     /**
@@ -496,8 +489,8 @@ public:
      * @param rRHSVariables parameter describing the expected RHSs
      */
     virtual void CalculateRightHandSide(std::vector< VectorType >& rRightHandSideVectors,
-					const std::vector< Variable< VectorType > >& rRHSVariables,
-					ProcessInfo& rCurrentProcessInfo)
+                    const std::vector< Variable< VectorType > >& rRHSVariables,
+                    ProcessInfo& rCurrentProcessInfo)
     {
     }
 
@@ -520,13 +513,13 @@ public:
      * @param rCurrentProcessInfo the current process info instance
      */
     virtual void CalculateFirstDerivativesContributions(MatrixType& rLeftHandSideMatrix,
-							VectorType& rRightHandSideVector,
-							ProcessInfo& rCurrentProcessInfo)
+                            VectorType& rRightHandSideVector,
+                            ProcessInfo& rCurrentProcessInfo)
     {
        if (rLeftHandSideMatrix.size1() != 0)
-	  rLeftHandSideMatrix.resize(0, 0, false);
+      rLeftHandSideMatrix.resize(0, 0, false);
         if (rRightHandSideVector.size() != 0)
-	  rRightHandSideVector.resize(0, false);
+      rRightHandSideVector.resize(0, false);
     }
 
     /**
@@ -536,10 +529,10 @@ public:
      * @param rCurrentProcessInfo the current process info instance
      */
     virtual void CalculateFirstDerivativesLHS(MatrixType& rLeftHandSideMatrix,
-					      ProcessInfo& rCurrentProcessInfo)
+                          ProcessInfo& rCurrentProcessInfo)
     {
         if (rLeftHandSideMatrix.size1() != 0)
-	  rLeftHandSideMatrix.resize(0, 0, false);
+      rLeftHandSideMatrix.resize(0, 0, false);
     }
 
 
@@ -550,10 +543,10 @@ public:
      * @param rCurrentProcessInfo the current process info instance
      */
     virtual void CalculateFirstDerivativesRHS(VectorType& rRightHandSideVector,
-					      ProcessInfo& rCurrentProcessInfo)
+                          ProcessInfo& rCurrentProcessInfo)
     {
         if (rRightHandSideVector.size() != 0)
-	  rRightHandSideVector.resize(0, false);
+      rRightHandSideVector.resize(0, false);
     }
 
 
@@ -576,13 +569,13 @@ public:
      * @param rCurrentProcessInfo the current process info instance
      */
     virtual void CalculateSecondDerivativesContributions(MatrixType& rLeftHandSideMatrix,
-							 VectorType& rRightHandSideVector,
-							 ProcessInfo& rCurrentProcessInfo)
+                             VectorType& rRightHandSideVector,
+                             ProcessInfo& rCurrentProcessInfo)
     {
        if (rLeftHandSideMatrix.size1() != 0)
-	  rLeftHandSideMatrix.resize(0, 0, false);
+      rLeftHandSideMatrix.resize(0, 0, false);
         if (rRightHandSideVector.size() != 0)
-	  rRightHandSideVector.resize(0, false);
+      rRightHandSideVector.resize(0, false);
     }
 
 
@@ -593,10 +586,10 @@ public:
      * @param rCurrentProcessInfo the current process info instance
      */
     virtual void CalculateSecondDerivativesLHS(MatrixType& rLeftHandSideMatrix,
-					       ProcessInfo& rCurrentProcessInfo)
+                           ProcessInfo& rCurrentProcessInfo)
     {
         if (rLeftHandSideMatrix.size1() != 0)
-	  rLeftHandSideMatrix.resize(0, 0, false);
+      rLeftHandSideMatrix.resize(0, 0, false);
     }
 
 
@@ -607,10 +600,10 @@ public:
      * @param rCurrentProcessInfo the current process info instance
      */
     virtual void CalculateSecondDerivativesRHS(VectorType& rRightHandSideVector,
-					       ProcessInfo& rCurrentProcessInfo)
+                           ProcessInfo& rCurrentProcessInfo)
     {
         if (rRightHandSideVector.size() != 0)
-	  rRightHandSideVector.resize(0, false);
+      rRightHandSideVector.resize(0, false);
     }
 
 
@@ -630,7 +623,7 @@ public:
     virtual void CalculateMassMatrix(MatrixType& rMassMatrix, ProcessInfo& rCurrentProcessInfo)
     {
         if (rMassMatrix.size1() != 0)
-	  rMassMatrix.resize(0, 0, false);
+      rMassMatrix.resize(0, 0, false);
     }
 
 
@@ -643,7 +636,7 @@ public:
     virtual void CalculateDampingMatrix(MatrixType& rDampingMatrix, ProcessInfo& rCurrentProcessInfo)
     {
         if (rDampingMatrix.size1() != 0)
-	  rDampingMatrix.resize(0, 0, false);
+      rDampingMatrix.resize(0, 0, false);
     }
 
 
@@ -763,44 +756,44 @@ public:
      */
 
     virtual void CalculateOnIntegrationPoints(const Variable<bool>& rVariable,
-					      std::vector<bool>& rOutput,
-					      const ProcessInfo& rCurrentProcessInfo)
+                          std::vector<bool>& rOutput,
+                          const ProcessInfo& rCurrentProcessInfo)
     {
     }
 
     virtual void CalculateOnIntegrationPoints(const Variable<int>& rVariable,
-					      std::vector<int>& rOutput,
-					      const ProcessInfo& rCurrentProcessInfo)
+                          std::vector<int>& rOutput,
+                          const ProcessInfo& rCurrentProcessInfo)
     {
     }
 
     virtual void CalculateOnIntegrationPoints(const Variable<double>& rVariable,
-					      std::vector<double>& rOutput,
-					      const ProcessInfo& rCurrentProcessInfo)
+                          std::vector<double>& rOutput,
+                          const ProcessInfo& rCurrentProcessInfo)
     {
     }
 
     virtual void CalculateOnIntegrationPoints(const Variable<array_1d<double, 3 > >& rVariable,
-					      std::vector< array_1d<double, 3 > >& rOutput,
-					      const ProcessInfo& rCurrentProcessInfo)
+                          std::vector< array_1d<double, 3 > >& rOutput,
+                          const ProcessInfo& rCurrentProcessInfo)
     {
     }
 
     virtual void CalculateOnIntegrationPoints(const Variable<array_1d<double, 6 > >& rVariable,
-					      std::vector< array_1d<double, 6 > >& rOutput,
-					      const ProcessInfo& rCurrentProcessInfo)
+                          std::vector< array_1d<double, 6 > >& rOutput,
+                          const ProcessInfo& rCurrentProcessInfo)
     {
     }
 
     virtual void CalculateOnIntegrationPoints(const Variable<Vector >& rVariable,
-					      std::vector< Vector >& rOutput,
-					      const ProcessInfo& rCurrentProcessInfo)
+                          std::vector< Vector >& rOutput,
+                          const ProcessInfo& rCurrentProcessInfo)
     {
     }
 
     virtual void CalculateOnIntegrationPoints(const Variable<Matrix >& rVariable,
-					      std::vector< Matrix >& rOutput,
-					      const ProcessInfo& rCurrentProcessInfo)
+                          std::vector< Matrix >& rOutput,
+                          const ProcessInfo& rCurrentProcessInfo)
     {
     }
 
@@ -817,99 +810,99 @@ public:
 
     //SET ON INTEGRATION POINTS - METHODS
     virtual void SetValueOnIntegrationPoints(const Variable<bool>& rVariable,
-					     std::vector<bool>& rValues,
-					     const ProcessInfo& rCurrentProcessInfo)
+                         std::vector<bool>& rValues,
+                         const ProcessInfo& rCurrentProcessInfo)
     {
     }
     virtual void SetValueOnIntegrationPoints(const Variable<int>& rVariable,
-					     std::vector<int>& rValues,
-					     const ProcessInfo& rCurrentProcessInfo)
+                         std::vector<int>& rValues,
+                         const ProcessInfo& rCurrentProcessInfo)
     {
     }
 
     virtual void SetValueOnIntegrationPoints(const Variable<double>& rVariable,
-					     std::vector<double>& rValues,
-					     const ProcessInfo& rCurrentProcessInfo)
+                         std::vector<double>& rValues,
+                         const ProcessInfo& rCurrentProcessInfo)
     {
     }
 
     virtual void SetValueOnIntegrationPoints(const Variable<array_1d<double, 3 > >& rVariable,
-					     std::vector<array_1d<double, 3 > > rValues,
-					     const ProcessInfo& rCurrentProcessInfo)
+                         std::vector<array_1d<double, 3 > > rValues,
+                         const ProcessInfo& rCurrentProcessInfo)
     {
     }
 
     virtual void SetValueOnIntegrationPoints(const Variable<array_1d<double, 6 > >& rVariable,
-					     std::vector<array_1d<double, 6 > > rValues,
-					     const ProcessInfo& rCurrentProcessInfo)
+                         std::vector<array_1d<double, 6 > > rValues,
+                         const ProcessInfo& rCurrentProcessInfo)
     {
     }
 
     virtual void SetValueOnIntegrationPoints(const Variable<Vector>& rVariable,
-					     std::vector<Vector>& rValues,
-					     const ProcessInfo& rCurrentProcessInfo)
+                         std::vector<Vector>& rValues,
+                         const ProcessInfo& rCurrentProcessInfo)
     {
     }
 
     virtual void SetValueOnIntegrationPoints(const Variable<Matrix>& rVariable,
-					     std::vector<Matrix>& rValues,
-					     const ProcessInfo& rCurrentProcessInfo)
+                         std::vector<Matrix>& rValues,
+                         const ProcessInfo& rCurrentProcessInfo)
     {
     }
 
     virtual void SetValueOnIntegrationPoints(const Variable<ConstitutiveLaw::Pointer>& rVariable,
-					     std::vector<ConstitutiveLaw::Pointer>& rValues,
-					     const ProcessInfo& rCurrentProcessInfo)
+                         std::vector<ConstitutiveLaw::Pointer>& rValues,
+                         const ProcessInfo& rCurrentProcessInfo)
     {
     }
 
     //GET ON INTEGRATION POINTS METHODS
 
     virtual void GetValueOnIntegrationPoints(const Variable<bool>& rVariable,
-					     std::vector<bool>& rValues,
-					     const ProcessInfo& rCurrentProcessInfo)
+                         std::vector<bool>& rValues,
+                         const ProcessInfo& rCurrentProcessInfo)
     {
     }
 
     virtual void GetValueOnIntegrationPoints(const Variable<int>& rVariable,
-					     std::vector<int>& rValues,
-					     const ProcessInfo& rCurrentProcessInfo)
+                         std::vector<int>& rValues,
+                         const ProcessInfo& rCurrentProcessInfo)
     {
     }
 
     virtual void GetValueOnIntegrationPoints(const Variable<double>& rVariable,
-					     std::vector<double>& rValues,
-					     const ProcessInfo& rCurrentProcessInfo)
+                         std::vector<double>& rValues,
+                         const ProcessInfo& rCurrentProcessInfo)
     {
     }
 
     virtual void GetValueOnIntegrationPoints(const Variable<array_1d<double, 3 > >& rVariable,
-					     std::vector<array_1d<double, 3 > >& rValues,
-					     const ProcessInfo& rCurrentProcessInfo)
+                         std::vector<array_1d<double, 3 > >& rValues,
+                         const ProcessInfo& rCurrentProcessInfo)
     {
     }
 
     virtual void GetValueOnIntegrationPoints(const Variable<array_1d<double, 6 > >& rVariable,
-					     std::vector<array_1d<double, 6 > >& rValues,
-					     const ProcessInfo& rCurrentProcessInfo)
+                         std::vector<array_1d<double, 6 > >& rValues,
+                         const ProcessInfo& rCurrentProcessInfo)
     {
     }
 
     virtual void GetValueOnIntegrationPoints(const Variable<Vector>& rVariable,
-					     std::vector<Vector>& rValues,
-					     const ProcessInfo& rCurrentProcessInfo)
+                         std::vector<Vector>& rValues,
+                         const ProcessInfo& rCurrentProcessInfo)
     {
     }
 
     virtual void GetValueOnIntegrationPoints(const Variable<Matrix>& rVariable,
-					     std::vector<Matrix>& rValues,
-					     const ProcessInfo& rCurrentProcessInfo)
+                         std::vector<Matrix>& rValues,
+                         const ProcessInfo& rCurrentProcessInfo)
     {
     }
 
     virtual void GetValueOnIntegrationPoints(const Variable<ConstitutiveLaw::Pointer>& rVariable,
-					     std::vector<ConstitutiveLaw::Pointer>& rValues,
-					     const ProcessInfo& rCurrentProcessInfo)
+                         std::vector<ConstitutiveLaw::Pointer>& rValues,
+                         const ProcessInfo& rCurrentProcessInfo)
     {
     }
 
@@ -946,7 +939,7 @@ public:
     virtual void MassMatrix(MatrixType& rMassMatrix, ProcessInfo& rCurrentProcessInfo)
     {
         if (rMassMatrix.size1() != 0)
-	  rMassMatrix.resize(0, 0, false);
+      rMassMatrix.resize(0, 0, false);
     }
 
     /**
@@ -969,7 +962,7 @@ public:
     virtual void DampMatrix(MatrixType& rDampMatrix, ProcessInfo& rCurrentProcessInfo)
     {
         if (rDampMatrix.size1() != 0)
-	  rDampMatrix.resize(0, 0, false);
+      rDampMatrix.resize(0, 0, false);
     }
 
     /**
@@ -991,7 +984,7 @@ public:
             VectorType& rRightHandSideVector, ProcessInfo& rCurrentProcessInfo)
     {
         if (rDampingMatrix.size1() != 0)
-	  rDampingMatrix.resize(0, 0, false);
+      rDampingMatrix.resize(0, 0, false);
 
     }
 
@@ -1120,25 +1113,6 @@ public:
         const TVariableType& rThisVariable) const
     {
         return mData.GetValue(rThisVariable);
-    }
-
-    ///@}
-    ///@name Flags
-    ///@{
-
-    Flags& GetFlags()
-      {
-	return *this;
-      }
-
-    Flags const& GetFlags() const
-    {
-      return *this;
-    }
-
-    void SetFlags(Flags const& rThisFlags)
-    {
-      Flags::operator=(rThisFlags);
     }
 
     ///@}
